Move guards to new header hpc/guards.h
