+++ src/cyclotom.c
@@ -119,6 +119,7 @@
 #include <src/saveload.h>               /* saving and loading */
 
 #include <src/code.h>
+#include <src/hpc/guards.h>
 
 /****************************************************************************
 **
