+++ src/hpc/aobjects.c
@@ -52,6 +52,7 @@
 
 #include <src/hpc/thread.h>
 #include <src/hpc/traverse.h>
+#include <src/hpc/guards.h>
 #include <src/vars.h>                   /* variables */
 
 #include <src/hpc/aobjects.h>
