+++ src/hpc/thread.c
@@ -13,6 +13,7 @@
 #include <src/stats.h>
 #include <src/gap.h>
 #include <src/hpc/misc.h>
+#include <src/hpc/guards.h>
 #include <src/hpc/thread.h>
 #include <src/hpc/threadapi.h>
 #include <src/fibhash.h>
