+++ src/hpc/traverse.c
@@ -14,6 +14,7 @@
 #include <src/precord.h>
 #include <src/stats.h>
 #include <src/gap.h>
+#include <src/hpc/guards.h>
 #include <src/hpc/thread.h>
 #include <src/hpc/traverse.h>
 #include <src/fibhash.h>
