+++ src/intrprtr.c
@@ -56,6 +56,7 @@
 #include <src/intrprtr.h>               /* interpreter */
 #include <src/intfuncs.h>
 
+#include <src/hpc/guards.h>
 #include <src/hpc/thread.h>
 #include <src/hpc/aobjects.h>           /* atomic objects */
