+++ src/objfgelm.h
@@ -11,7 +11,7 @@
 #ifndef GAP_OBJFGELM_H
 #define GAP_OBJFGELM_H
 
-#include <src/hpc/tls.h>
+#include <src/hpc/guards.h>
 
 /****************************************************************************
 **
