+++ src/opers.c
@@ -50,6 +50,7 @@
 #include <src/gmpints.h>   
 
 #ifdef HPCGAP
+#include <src/hpc/guards.h>
 #include <src/hpc/thread.h>             /* threads */
 #include <src/hpc/aobjects.h>           /* atomic objects */
