+++ src/permutat.c
@@ -71,6 +71,7 @@
 #include <src/code.h>                   /* coder */
 
 #include <src/saveload.h>               /* saving and loading */
+#include <src/hpc/guards.h>
 
 #include <src/trans.h>
 #include <assert.h>
