+++ src/plist.c
@@ -67,6 +67,7 @@
 
 #include <src/saveload.h>               /* saving and loading */
 #include <src/code.h>
+#include <src/hpc/guards.h>
 
 #include <src/gaputils.h>
