+++ src/pperm.c
@@ -46,6 +46,7 @@
 #include <src/set.h>                    /* sets */
 
 #include <src/code.h>                   /* coder */
+#include <src/hpc/guards.h>
 
 #include <src/pperm.h>                  /* same header applies */
