+++ src/precord.c
@@ -57,6 +57,7 @@
 #include <src/saveload.h>               /* saving and loading */
 
 #include <src/code.h>                   /* coder */
+#include <src/hpc/guards.h>
 #include <src/hpc/aobjects.h>           /* thread-local storage */
 
 #include <src/gaputils.h>
