+++ src/range.c
@@ -79,6 +79,7 @@
 #include <src/saveload.h>               /* saving and loading */
 
 #include <src/code.h>                   /* coder */
+#include <src/hpc/guards.h>
 
 #include <src/gaputils.h>
