+++ src/weakptr.c
@@ -39,6 +39,7 @@
 
 #include <src/scanner.h>                /* scanner */
 #include <src/code.h>                   /* coder */
+#include <src/hpc/guards.h>
 
 #ifdef BOEHM_GC
 # ifdef HPCGAP
